+++ examples/BUILD
@@ -39,10 +39,5 @@ bazel_integration_test(
 )
 
 bazel_integration_test(
-    name = "relative_requirements_example",
-    timeout = "long",
-)
-
-bazel_integration_test(
     name = "bzlmod_example",
 )
