+++ examples/pip_repository_annotations/WORKSPACE
@@ -54,9 +54,9 @@ pip_parse(
     requirements_lock = "//:requirements.txt",
 )
 
-load("@pip_parsed//:requirements.bzl", "install_deps")
+load("@pip_parsed//:requirements.bzl", install_pip_parse_deps = "install_deps")
 
-install_deps()
+install_pip_parse_deps()
 
 # For a more thorough example of `pip_install`. See `@rules_python//examples/pip_install`
 pip_install(
@@ -65,3 +65,7 @@ pip_install(
     python_interpreter_target = interpreter,
     requirements = "//:requirements.txt",
 )
+
+load("@pip_installed//:requirements.bzl", install_pip_install_deps = "install_deps")
+
+install_pip_install_deps()
