+++ examples/pip_repository_annotations/requirements.txt
@@ -9,4 +9,4 @@
 wheel==0.37.1 \
     --hash=sha256:4bdcd7d840138086126cd09254dc6195fb4fc6f01c050a1d7236f2630db1d22a \
     --hash=sha256:e9a504e793efbca1b8e0e9cb979a249cf4a0a7b5b8c9e8b65a5e39d49529c1c4
-    # via -r requirements.in
+    # via -r ./requirements.in
