--- examples/relative_requirements/WORKSPACE
@@ -1,21 +1,0 @@
-workspace(name = "example_repo")
-
-local_repository(
-    name = "rules_python",
-    path = "../..",
-)
-
-load("@rules_python//python:repositories.bzl", "python_register_toolchains")
-
-python_register_toolchains(
-    name = "python39",
-    python_version = "3.9",
-)
-
-load("@python39//:defs.bzl", "interpreter")
-load("@rules_python//python:pip.bzl", "pip_install")
-
-pip_install(
-    python_interpreter_target = interpreter,
-    requirements = "//:requirements.txt",
-)
