--- examples/relative_requirements/main.py
@@ -1,5 +1,0 @@
-import relative_package_name
-
-if __name__ == "__main__":
-    # Run a function from the relative package
-    print(relative_package_name.test())
