--- examples/relative_requirements/relative_package/relative_package_name/__init__.py
@@ -1,2 +1,0 @@
-def test():
-    return True
