--- examples/relative_requirements/relative_package/setup.py
@@ -1,7 +1,0 @@
-from setuptools import setup
-
-setup(
-    name="relative_package_name",
-    version="1.0.0",
-    packages=["relative_package_name"],
-)
