--- examples/relative_requirements/requirements.txt
@@ -1,1 +1,0 @@
-./relative_package
