+++ python/pip_install/extract_wheels/arguments.py
@@ -39,12 +39,6 @@ def parse_common_args(parser: ArgumentParser) -> ArgumentParser:
         required=True,
         help="Prefix to prepend to packages",
     )
-    parser.add_argument(
-        "--download_only",
-        action="store_true",
-        help="Use 'pip download' instead of 'pip wheel'. Disables building wheels from source, but allows use of "
-             "--platform, --python-version, --implementation, and --abi in --extra_pip_args.",
-    )
     return parser
