--- tests/pip_deps/README.md
@@ -1,1 +1,0 @@
-Run `bazel run @unpinned_pip//:pin.update` to keep `PIP_PACKAGES` in `//:WORKSPACE` in sync with `//:requirements.txt`.
