--- tests/pip_deps/WORKSPACE
@@ -1,26 +1,0 @@
-workspace(name = "pip_deps")
-
-local_repository(
-    name = "rules_python",
-    path = "../..",
-)
-
-load("@rules_python//python:repositories.bzl", "python_register_toolchains")
-
-# This toolchain is explicitly 3.10 while `rules_python` is 3.9 to act as
-# a regression test, ensuring 3.10 is functional
-python_register_toolchains(
-    name = "python310",
-    python_version = "3.10",
-)
-
-load("@python310//:defs.bzl", "interpreter")
-load("//:pip_deps.bzl", "pip_deps")
-
-PIP_PACKAGES = {"sphinx": "4.5.0"}
-
-pip_deps(
-    name = "pip",
-    packages = PIP_PACKAGES,
-    python_interpreter_target = interpreter,
-)
