--- tests/pip_deps/test.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-set -euo pipefail
-
-! grep external "$1"
