+++ tests/pip_repository_entry_points/WORKSPACE
@@ -24,9 +24,9 @@ pip_parse(
     requirements_lock = "//:requirements.txt",
 )
 
-load("@pip_parsed//:requirements.bzl", "install_deps")
+load("@pip_parsed//:requirements.bzl", install_pip_parse_deps = "install_deps")
 
-install_deps()
+install_pip_parse_deps()
 
 # For a more thorough example of `pip_install`. See `@rules_python//examples/pip_install`
 pip_install(
@@ -34,3 +34,7 @@ pip_install(
     python_interpreter_target = interpreter,
     requirements = "//:requirements.txt",
 )
+
+load("@pip_installed//:requirements.bzl", install_pip_install_deps = "install_deps")
+
+install_pip_install_deps()
